Move CodeSchemes to its own file
