+++ src/lib/pipeline_configuration.py
@@ -5,6 +5,8 @@ from core_data_modules.cleaners import somali, Codes
 from core_data_modules.data_models import Scheme, validators
 from dateutil.parser import isoparse
 
+from src.lib.code_schemes import CodeSchemes
+
 
 class CodingModes(object):
     SINGLE = "SINGLE"
@@ -16,25 +18,6 @@ class FoldingModes(object):
     YES_NO_AMB = "YES_NO_AMB"
     CONCATENATE = "CONCATENATE"
     MATRIX = "MATRIX"
-
-
-def _open_scheme(filename):
-    with open(f"code_schemes/{filename}", "r") as f:
-        firebase_map = json.load(f)
-        return Scheme.from_firebase_map(firebase_map)
-
-
-class CodeSchemes(object):
-    SOMALIA_OPERATOR = _open_scheme("somalia_operator.json")
-
-    S03E01_BOSSASO_REASONS = _open_scheme("s03e01_bossaso_reasons.json")
-    S03E02_BOSSASO_REASONS = _open_scheme("s03e02_bossaso_reasons.json")
-    S03E03_BOSSASO_REASONS = _open_scheme("s03e03_bossaso_reasons.json")
-    S03E04_BOSSASO_REASONS = _open_scheme("s03e04_bossaso_reasons.json")
-
-    AGE = _open_scheme("age.json")
-    GENDER = _open_scheme("gender.json")
-    RECENTLY_DISPLACED = _open_scheme("recently_displaced.json")
 
 
 class CodingConfiguration(object):
